Moved "Substring" methods to a separate class "Selector".
Created test class "SelectorTest".
